+++ std_xyx/std_xyx/start_menue_scene.h
@@ -11,7 +11,7 @@ public:
     ~StartMenueScene();
 
     void on_enter()override;
-    void on_update(int delta)override;
+    void on_update(float delta)override;
     void on_input(const SDL_Event & event)override;
     void on_render()override;
     void on_exit()override;
